+++ src/Models/Register.php
@@ -1,13 +1,13 @@
 <?php namespace Experiensa\Plugin\Models;
 /* Custom Post Types */
-use Experiensa\Plugin\Models\PostType\Brochure;
+// use Experiensa\Plugin\Models\PostType\Brochure;
 use Experiensa\Plugin\Models\PostType\Estimate;
 use Experiensa\Plugin\Models\PostType\Facility;
 use Experiensa\Plugin\Models\PostType\Feedback;
 use Experiensa\Plugin\Models\PostType\Host;
 use Experiensa\Plugin\Models\PostType\Partner;
 use Experiensa\Plugin\Models\PostType\Place;
-use Experiensa\Plugin\Models\PostType\Service;
+// use Experiensa\Plugin\Models\PostType\Service;
 use Experiensa\Plugin\Models\PostType\Voyage;
 /* Taxonomies */
 use Experiensa\Plugin\Models\Taxonomy\Country;
@@ -24,14 +24,14 @@ class Register{
         /**
          * Post types
          */
-        Brochure::addCustomPostType();
+        // Brochure::addCustomPostType();
         Estimate::addCustomPostType();
         Facility::addCustomPostType();
         Feedback::addCustomPostType();
         Host::addCustomPostType();
         Partner::addCustomPostType();
         Place::addCustomPostType();
-        Service::addCustomPostType();
+        // Service::addCustomPostType();
         Voyage::addCustomPostType();
         /**
          * Taxonomies
@@ -44,7 +44,6 @@ class Register{
         Location::addTaxonomy();
         ProductType::addTaxonomy();
         Theme::addTaxonomy();
-        // register_activation_hook( EXPERIENSA_FILE, array( __CLASS__, 'register_flush_rewrite_rules') );
     }
     public static function register_flush_rewrite_rules(){
         self::init();