+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.9.4: May 26th, 2018
+* Brochure, Services post type desactivated
+* Voyage Flight, Itinerary, Accommodation desactivated
+* Voyage metaboxes name changed
 ### 0.9.4: April 29th, 2018
 * 'Regions' taxonomy added ([6af89e8](https://github.com/Experiensa/wp-experiensa/commit/6af89e8717e87f3fb1829c3fe9b0a52e9515a69a))
 * Change catalog label titles to french language ([6af89e8](https://github.com/Experiensa/wp-experiensa/commit/6af89e8717e87f3fb1829c3fe9b0a52e9515a69a))